Add draw offset to ObjectDrawInfo

+++ vulkan-tutorial/graphics/texture_atlas.h
@@ -27,8 +27,8 @@ struct TextureInfo
 
 struct DrawOffset
 {
-	int8_t x;
-	int8_t y;
+	int x;
+	int y;
 };
 
 struct TextureAtlas
